+++ node/utils/GraphQLError.ts
@@ -3,6 +3,14 @@ export default class GraphQLError extends Error {
 
   constructor(message: string, details?: any) {
     super(message)
-    this.extensions = { message, ...details }
+    this.extensions = {
+      message,
+      ...details,
+      exception: {
+        ...details?.exception,
+        // vtex-node-api checks err?.extensions?.exception?.level before logging
+        level: details?.logLevel || 'error',
+      },
+    }
   }
 }